# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customer/profile": {
      "get": {
        "operationId": "getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "CustomerProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "CustomerProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "createProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "CustomerProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/customer/profile/fcm-token": {
      "put": {
        "operationId": "upsertFcmToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsertFcmTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "CustomerProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/driver/profile": {
      "get": {
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "includeDocuments",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "includeVehicle",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "createProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DriverProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/driver/profile/status": {
      "put": {
        "operationId": "updateDriverStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/driver/profile/fcm-token": {
      "put": {
        "operationId": "upsertFcmToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsertFcmTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/driver/profile/payout-details": {
      "put": {
        "operationId": "updateDriverPayoutDetails",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePayoutDetailsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverProfile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/driver/address": {
      "get": {
        "operationId": "getAddress",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "DriverAddress"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "createAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "DriverAddress"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "updateAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "DriverAddress"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteAddress",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "DriverAddress"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Hello Truck API",
    "description": "API documentation for Hello Truck application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your access token",
        "in": "header"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "CreateGstDetailsDto": {
        "type": "object",
        "properties": {
          "gstNumber": {
            "type": "string",
            "pattern": "/^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$/"
          },
          "businessName": {
            "type": "string"
          },
          "businessAddress": {
            "type": "string"
          }
        },
        "required": [
          "gstNumber",
          "businessName",
          "businessAddress"
        ]
      },
...
      "GetProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "isBusiness": {
            "type": "boolean"
          },
          "referralCode": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isActive",
          "firstName",
          "lastName",
          "email",
          "isBusiness",
          "referralCode",
          "phoneNumber",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "googleIdToken": {
            "type": "string"
          }
        }
      },
      "i": {
        "type": "object",
        "properties": {}
      },
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "formattedAddress": {
            "type": "string"
          },
          "addressDetails": {
            "type": "string"
          },
          "latitude": {
            "$ref": "#/components/schemas/i"
          },
          "longitude": {
            "$ref": "#/components/schemas/i"
          }
        },
        "required": [
          "formattedAddress",
          "latitude",
          "longitude"
        ]
      },
      "CreateSavedAddressDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "noteToDriver": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "address": {
            "$ref": "#/components/schemas/CreateAddressDto"
          }
        },
        "required": [
          "name",
          "contactName",
          "contactPhone",
          "address"
        ]
      },
      "CreateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "googleIdToken": {
            "type": "string"
          },
          "referralCode": {
            "type": "string"
          },
          "gstDetails": {
            "$ref": "#/components/schemas/CreateGstDetailsDto"
          },
          "savedAddress": {
            "$ref": "#/components/schemas/CreateSavedAddressDto"
          }
        },
        "required": [
          "firstName"
        ]
      },
      "UsertFcmTokenDto": {
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string"
          }
        },
        "required": [
          "fcmToken"
        ]
      },
...
      "CreateDriverDocumentsDto": {
        "type": "object",
        "properties": {
          "licenseUrl": {
            "type": "string",
            "format": "uri"
          },
          "rcBookUrl": {
            "type": "string",
            "format": "uri"
          },
          "fcUrl": {
            "type": "string",
            "format": "uri"
          },
          "insuranceUrl": {
            "type": "string",
            "format": "uri"
          },
          "aadharUrl": {
            "type": "string",
            "format": "uri"
          },
          "panNumber": {
            "type": "string",
            "pattern": "/^[A-Z]{5}[0-9]{4}[A-Z]{1}$/"
          },
          "ebBillUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "licenseUrl",
          "rcBookUrl",
          "fcUrl",
          "insuranceUrl",
          "aadharUrl",
          "panNumber",
          "ebBillUrl"
        ]
      },
      "CreateDriverAddressDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string"
          },
          "landmark": {
            "type": "string"
          },
          "pincode": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "pattern": "/^[0-9]{6}$/"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "latitude": {
            "$ref": "#/components/schemas/i"
          },
          "longitude": {
            "$ref": "#/components/schemas/i"
          }
        },
        "required": [
          "addressLine1",
          "pincode",
          "city",
          "district",
          "state"
        ]
      },
      "CreateVehicleOwnerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "aadharUrl": {
            "type": "string",
            "format": "uri"
          },
          "contactNumber": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "landmark": {
            "type": "string"
          },
          "pincode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "aadharUrl",
          "contactNumber",
          "addressLine1",
          "pincode",
          "city",
          "district",
          "state"
        ]
      },
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "vehicleNumber": {
            "type": "string"
          },
          "vehicleType": {
            "type": "object"
          },
          "vehicleBodyLength": {
            "minimum": 1,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "vehicleBodyType": {
            "type": "object"
          },
          "fuelType": {
            "type": "object"
          },
          "vehicleImageUrl": {
            "type": "string",
            "format": "uri"
          },
          "owner": {
            "$ref": "#/components/schemas/CreateVehicleOwnerDto"
          }
        },
        "required": [
          "vehicleNumber",
          "vehicleType",
          "vehicleBodyLength",
          "vehicleBodyType",
          "fuelType",
          "vehicleImageUrl"
        ]
      },
      "CreateBankDetailsDto": {
        "type": "object",
        "properties": {
          "accountHolderName": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "ifscCode": {
            "type": "string",
            "pattern": "/^[A-Z]{4}0[A-Z0-9]{6}$/"
          }
        },
        "required": [
          "accountHolderName",
          "accountNumber",
          "ifscCode"
        ]
      },
      "CreateVpaDetailsDto": {
        "type": "object",
        "properties": {
          "vpa": {
            "type": "string",
            "pattern": "/^[a-zA-Z0-9\\.\\-_]{2,256}@[a-zA-Z0-9\\.\\-_]{2,64}$/"
          }
        },
        "required": [
          "vpa"
        ]
      },
      "CreatePayoutDetailsDto": {
        "type": "object",
        "properties": {
          "payoutMethod": {
            "type": "string",
            "enum": [
              "BANK_ACCOUNT",
              "VPA"
            ]
          },
          "bankDetails": {
            "$ref": "#/components/schemas/CreateBankDetailsDto"
          },
          "vpaDetails": {
            "$ref": "#/components/schemas/CreateVpaDetailsDto"
          }
        },
        "required": [
          "payoutMethod"
        ]
      },
      "CreateDriverProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "googleIdToken": {
            "type": "string"
          },
          "alternatePhone": {
            "type": "string"
          },
          "referalCode": {
            "type": "string"
          },
          "photo": {
            "type": "string",
            "format": "uri"
          },
          "documents": {
            "$ref": "#/components/schemas/CreateDriverDocumentsDto"
          },
          "address": {
            "$ref": "#/components/schemas/CreateDriverAddressDto"
          },
          "vehicle": {
            "$ref": "#/components/schemas/CreateVehicleDto"
          },
          "payoutDetails": {
            "$ref": "#/components/schemas/CreatePayoutDetailsDto"
          }
        },
        "required": [
          "firstName",
          "documents",
          "address",
          "vehicle",
          "payoutDetails"
        ]
      },
      "UpdatePayoutDetailsDto": {
        "type": "object",
        "properties": {
          "payoutMethod": {
            "enum": [
              "BANK_ACCOUNT",
              "VPA"
            ],
            "type": "string"
          },
          "bankDetails": {
            "$ref": "#/components/schemas/CreateBankDetailsDto"
          },
          "vpaDetails": {
            "$ref": "#/components/schemas/CreateVpaDetailsDto"
          }
        },
        "required": [
          "payoutMethod"
        ]
      },
      "UpdateDriverProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "googleIdToken": {
            "type": "string"
          },
          "alternatePhone": {
            "type": "string"
          },
          "photo": {
            "type": "string",
            "format": "uri"
          },
          "payoutDetails": {
            "$ref": "#/components/schemas/UpdatePayoutDetailsDto"
          }
        }
      },
      "UpdateDriverStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "AddressResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "driverId": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string"
          },
          "landmark": {
            "type": "string",
            "nullable": true
          },
          "pincode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "latitude": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "longitude": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "driverId",
          "addressLine1",
          "landmark",
          "pincode",
          "city",
          "district",
          "state",
          "latitude",
          "longitude",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateDriverAddressDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string"
          },
          "landmark": {
            "type": "string"
          },
          "pincode": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "pattern": "/^[0-9]{6}$/"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "latitude": {
            "$ref": "#/components/schemas/i"
          },
          "longitude": {
            "$ref": "#/components/schemas/i"
          }
        }
      }
    }
  }
}